% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csrhub_keygen.R
\name{csrhub_keygen}
\alias{csrhub_keygen}
\title{CSRHub Generate Session ID}
\arguments{
\item{email}{csrhub email associated with your account}

\item{password}{csrhub email for associated account}
}
\description{
This function creates a session ID to use the API. It will expire within 24 hours.
}
\keyword{id,}
\keyword{key}
\keyword{session,}
