% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAlias.R
\name{GetAlias}
\alias{GetAlias}
\title{CARHUB Search for Company API Alias}
\usage{
GetAlias(company, key)
}
\arguments{
\item{company}{pass a company name or multiple company names for alias match}

\item{key}{use the key generated with the csrhub_keygen() function}
}
\description{
The function will recieve the csrhub api alias and return a data frame with searched term
for the company you search for. Currently only written to work with names but other
searches are possible by ISIN or Ticker.
}
\examples{
company_names <- c("apple","alphabet")
results <- GetAlias(company=company_names,key=key)
print(results)
search_term    CSRNAME
1       apple  Apple-Inc
2    alphabet Google-Inc

}
\keyword{alias,}
\keyword{search}
